requirements.txt
django==4.2.4
django-cors-headers==4.2.0
djangorestframework  #==3.14.0
djangorestframework-simplejwt #==5.3.0
drf-yasg==1.21.7
django-filter==23.2
requests==2.31.0
python-dotenv==1.0.0
mysqlclient
reportlab==4.0.7
xhtml2pdf==0.2.13s